Fix file conflicts
